+++ src/negocio/SAUsuario.java
@@ -8,6 +8,6 @@ public interface SAUsuario {
     public Collection<TUsuario> readAll();
     public void update(TUsuario tUsuario);
     public void delete (int id);
-    public void actualizarSaldo(int cantidad);
+    public void actualizarSaldo(double cantidad);
     public void actualizarSuscr(int id);
 }
