+++ src/negocio/Usuario.java
@@ -14,13 +14,13 @@ public class Usuario {
     protected String pais;
     protected Suscripciones suscripcion;
     protected String direccion;
-    protected int saldo;
+    protected double saldo;
     protected int id;
     protected int id_cesta;
     protected List<Integer> favoritos;//tiene una lista de los id de los articulos favoritos
 
 
-    protected Usuario(int ID, String Nombre, String Apellidos, String Correo, String Contrasenya, int anyoNac, char Sexo, String Pais, Suscripciones Suscripcion, String Direccion, int Saldo, int idCesta) {
+    protected Usuario(int ID, String Nombre, String Apellidos, String Correo, String Contrasenya, int anyoNac, char Sexo, String Pais, Suscripciones Suscripcion, String Direccion, double Saldo, int idCesta) {
         id = ID;
         nombre = Nombre;
         apellidos = Apellidos;
@@ -106,11 +106,11 @@ public class Usuario {
         this.direccion = direccion;
     }
 
-    public int getSaldo() {
+    public double getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(int saldo) {
+    public void setSaldo(double saldo) {
         this.saldo = saldo;
     }
